+++ src/tests/blueprint/python/t_python_blueprint_mesh.py
@@ -252,6 +252,18 @@ class Test_Blueprint_Mesh(unittest.TestCase):
             blueprint.mesh.examples.generate(ename,n)
             self.assertTrue(blueprint.mesh.verify(n,info))
 
+
+    def test_remove(self):
+        n    = conduit.Node()
+        opts = conduit.Node()
+        info = conduit.Node()
+        blueprint.mesh.examples.generate("braid",n)
+        opts["fields"]= ["braid","vel"]
+        blueprint.mesh.remove(opts,n)
+        self.assertFalse(n.has_path("fields/braid"))
+        self.assertFalse(n.has_path("fields/vel"))
+        self.assertTrue(n.has_path("fields/radial"))
+
     def test_partition(self):
         def same(a, b):
             for i in range(len(a)):